+++ Cmdlets/Public/Invoke-TeamViewerPackageDownload.ps1
@@ -1,5 +1,7 @@
 function Invoke-TeamViewerPackageDownload {
-    Param(
+    [CmdletBinding()]
+
+    param(
         [Parameter()]
         [ValidateSet('Full', 'Host', 'MSI32', 'MSI64', 'Portable', 'QuickJoin', 'QuickSupport', 'Full64Bit')]
         [string]
@@ -7,15 +9,15 @@ function Invoke-TeamViewerPackageDownload {
 
         [Parameter()]
         [ValidateScript( {
-                if($PackageType -eq 'MSI32' -or 'MSI64' ){
+                if ($PackageType -eq 'MSI32' -or 'MSI64' ) {
                     $PSCmdlet.ThrowTerminatingError(
-                        ("MajorVersion parameter is not supported for MSI packages" | `
-                                ConvertTo-ErrorRecord -ErrorCategory InvalidArgument))
+                        ('MajorVersion parameter is not supported for MSI packages' | `
+                            ConvertTo-ErrorRecord -ErrorCategory InvalidArgument))
                 }
                 if ($_ -lt 14) {
                     $PSCmdlet.ThrowTerminatingError(
                         ("Unsupported TeamViewer version $_" | `
-                                ConvertTo-ErrorRecord -ErrorCategory InvalidArgument))
+                            ConvertTo-ErrorRecord -ErrorCategory InvalidArgument))
                 }
                 $true
             } )]
@@ -39,32 +41,52 @@ function Invoke-TeamViewerPackageDownload {
 
     $additionalPath = ''
     $filename = switch ($PackageType) {
-        'Full' { 'TeamViewer_Setup.exe' }
-        'MSI32' { 'TeamViewer_MSI32.zip' }
-        'MSI64' { 'TeamViewer_MSI64.zip' }
-        'Host' { 'TeamViewer_Host_Setup.exe' }
-        'Portable' { 'TeamViewerPortable.zip' }
-        'QuickJoin' { 'TeamViewerQJ.exe' }
-        'QuickSupport' { 'TeamViewerQS.exe' }
-        'Full64Bit' { 'TeamViewer_Setup_x64.exe' }
+        'Full' {
+            'TeamViewer_Setup.exe'
+        }
+        'MSI32' {
+            'TeamViewer_MSI32.zip'
+        }
+        'MSI64' {
+            'TeamViewer_MSI64.zip'
+        }
+        'Host' {
+            'TeamViewer_Host_Setup.exe'
+        }
+        'Portable' {
+            'TeamViewerPortable.zip'
+        }
+        'QuickJoin' {
+            'TeamViewerQJ.exe'
+        }
+        'QuickSupport' {
+            'TeamViewerQS.exe'
+        }
+        'Full64Bit' {
+            'TeamViewer_Setup_x64.exe'
+        }
     }
+
     if ($MajorVersion) {
         $additionalPath = "/version_$($MajorVersion)x"
     }
-    if(($PackageType -eq 'MSI32' -or 'MSI64' )){
+
+    if (($PackageType -eq 'MSI32' -or 'MSI64' )) {
         $additionalPath = '/version_15x'
     }
 
     $downloadUrl = "https://dl.teamviewer.com/download$additionalPath/$filename"
     $targetFile = Join-Path $TargetDirectory $filename
 
-    if ((Test-Path $targetFile) -And -Not $Force -And `
-            -Not $PSCmdlet.ShouldContinue("File $targetFile already exists. Override?", "Override existing file?")) {
+    if ((Test-Path $targetFile) -and -not $Force -and `
+            -not $PSCmdlet.ShouldContinue("File $targetFile already exists. Override?", 'Override existing file?')) {
         return
     }
 
     Write-Verbose "Downloading $downloadUrl to $targetFile"
+
     $client = New-Object System.Net.WebClient
     $client.DownloadFile($downloadUrl, $targetFile)
+
     Write-Output $targetFile
 }