+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Adds `CmdletBinding` to public commands.
 - Standardizes pipeline emission on Write-Output.
 - Creates, expands, and hardens all private functions tests.
 - Automate release module versioning.